+++ docs/release_notes.md
@@ -1,7 +1,7 @@
 # Release Notes
 
 ## unreleased
-* __Template类中新增layout参数，以支持排版布局的灵活配置。__
+* __Template类中新增layout参数，以支持排版布局的灵活配置。首期支持：GRID（网格布局），FLOW（流动布局）。__
 * __Template类移除font_size参数，字体大小改为使用font参数的size属性表示。__
 
 ## v6.0.0 (2020-10-18)